+++ app/assets/stylesheets/address.scss
@@ -1,0 +1,3 @@
+// Place all the styles related to the address controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: http://sass-lang.com/
