+++ app/helpers/address_helper.rb
@@ -1,0 +1,2 @@
+module AddressHelper
+end
