+++ app/views/users/show.html.erb
@@ -68,12 +68,12 @@
             <h2 class="content__mypage__left__side__list__menu">設定</h2>
             <ul>
               <li>
-                <%= link_to profile_user_path, class: "content__mypage__left__side__list__menu__tag" do %>
+                <%= link_to address_path, class: "content__mypage__left__side__list__menu__tag" do %>
                   プロフィール<%= fa_icon("angle-right") %>
                 <% end %>
               </li>
               <li>
-                <%= link_to edit_user_path, class: "content__mypage__left__side__list__menu__tag" do %>
+                <%= link_to edit_address_path, class: "content__mypage__left__side__list__menu__tag" do %>
                   住所変更<%= fa_icon("angle-right") %>
                 <% end %>
               </li>
@@ -83,7 +83,7 @@
                 <% end %>
               </li>
               <li>
-                <%= link_to destroy_user_session_path, class: "content__mypage__left__side__list__menu__tag" do %>
+                <%= link_to destroy_user_session_path, method: :delete, class: "content__mypage__left__side__list__menu__tag" do %>
                   ログアウト<%= fa_icon("angle-right") %>
                 <% end %>
               </li>
@@ -91,9 +91,9 @@
             <h2 class="content__mypage__left__side__list__menu">退会</h2>
             <ul>
               <li>
-                <%= link_to user_path(@user.id), method: :delete, data: { confirm: '本当に削除しますか?' }, class: "content__mypage__left__side__list__menu__tag" do %>
-                  退会<%= fa_icon("angle-right") %>
-                <% end %>
+                <%# <%= link_to user_path(@user.id), method: :delete, data: { confirm: '本当に削除しますか?' }, class: "content__mypage__left__side__list__menu__tag" do %>
+                  <%# 退会<%= fa_icon("angle-right") %>
+                <%# <% end %>
               </li>
             </ul>
           </li>
