+++ test/controllers/address_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class AddressControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
